refactor: move filtering and patching code to user_repository
