--- internal/api/scim/patch.go
@@ -1,1 +1,0 @@
-package scim
